// world-cup-api/src/match/model/match-action.enum.ts
export enum MatchAction {
  RESTART_MATCH = 'RESTART_MATCH',
  ATTACK = 'ATTACK',
  HOLD = 'HOLD',
  PRESS = 'PRESS',
  LONG_PASS = 'LONG_PASS',
  SHOOT = 'SHOOT',
  PASS = 'PASS',
  DRIBBLE = 'DRIBBLE',
  CROSS = 'CROSS',
  DEFEND = 'DEFEND',
  LEFT = 'LEFT',
  RIGHT = 'RIGHT',
  CENTER = 'CENTER',
  PICAR = 'PICAR',
  DIVE_LEFT = 'DIVE_LEFT',
  DIVE_RIGHT = 'DIVE_RIGHT',
  STAY_CENTER = 'STAY_CENTER',
  WAIT = 'WAIT',
  BLOCK = 'BLOCK',
  TACKLE = 'TACKLE',
  QUIT_MATCH = 'QUIT_MATCH',
}