// world-cup-api/src/match/model/match-action-outcome.enum.ts
export enum MatchActionOutcome {
  PASS_SUCCESS_PROGRESS = 'PASS_SUCCESS_PROGRESS',
  PASS_INTERCEPTED = 'PASS_INTERCEPTED',
  LONG_PASS_SUCCESS_PROGRESS = 'LONG_PASS_SUCCESS_PROGRESS',
  LONG_PASS_LOST = 'LONG_PASS_LOST',
  DRIBBLE_WON = 'DRIBBLE_WON',
  DRIBBLE_LOST = 'DRIBBLE_LOST',
  PRESS_WON = 'PRESS_WON',
  PRESS_LOST = 'PRESS_LOST',
  TACKLE_WON = 'TACKLE_WON',
  TACKLE_LOST = 'TACKLE_LOST',
  DEFEND_HOLD = 'DEFEND_HOLD',
  DEFEND_BROKEN = 'DEFEND_BROKEN',
  ATTACK_PROGRESS = 'ATTACK_PROGRESS',
  ATTACK_STALLED = 'ATTACK_STALLED',
  HOLD_STABLE = 'HOLD_STABLE',
  HOLD_LOST = 'HOLD_LOST',
  CROSS_CONNECTED = 'CROSS_CONNECTED',
  CROSS_CLEARED = 'CROSS_CLEARED',
  SHOOT_USER_GOAL = 'SHOOT_USER_GOAL',
  SHOOT_OPPONENT_GOAL = 'SHOOT_OPPONENT_GOAL',
  SHOOT_BLOCKED_REBOUND_FOR = 'SHOOT_BLOCKED_REBOUND_FOR',
  SHOOT_BLOCKED_REBOUND_AGAINST = 'SHOOT_BLOCKED_REBOUND_AGAINST',
  SHOOT_SAVED_REBOUND_FOR = 'SHOOT_SAVED_REBOUND_FOR',
  SHOOT_SAVED_REBOUND_AGAINST = 'SHOOT_SAVED_REBOUND_AGAINST',
  SHOOT_MISSED = 'SHOOT_MISSED',
  NOT_HANDLED = 'NOT_HANDLED',
}